refactor: extract OpenAPI spec into shared module
